use interface service
